refactor(hook): the emulator half moves to files a deck never compiles
